style: include order, alignment, type narrowing in CompactedRegistries

+++ src/registries/CompactedRegistries.cpp
@@ -59,14 +59,12 @@ Item &EnchReg::refresh_item(Item &item) const {
     uint16_t total_lsum = 0;
     bool is_book = (item.type == ItemType::Book);
 
-    for (const auto& ench : item.enchs) {
-        const auto& info_mask = (*this)[ench.id].exc_mask;
+    for (const auto &ench : item.enchs) {
+        const auto &info_mask = (*this)[ench.id].exc_mask;
         for (size_t k = 0; k < _mask_size; ++k)
             item.exc_mask[k] |= info_mask[k];
 
-        uint16_t mult = is_book
-            ? std::max(1, get_multiplier(ench.id) >> 1)
-            : get_multiplier(ench.id);
+        uint16_t mult = is_book ? static_cast<uint16_t>(std::max(1, get_multiplier(ench.id) >> 1)) : get_multiplier(ench.id);
         total_lsum += ench.level * mult;
     }
